Move to resetting the sensor in hardware The step counter is now set in hardware to the previous reading + any changes when the daemon starts This means that the step counter now just records the raw step counter value, and adjusts it correctly when needed
This introduces a bug where the step counter will show bad readings if the daemon is restarted without a system reboot:
the step counter's current value will be added to the last recorded value, which will cause the number of steps to effectively double in most situations.
This might be fixed with some code to detect if a system reboot has passed since the last time the counter started.